Fixes #113 - Renamed errorHandling package to errorHandler
